change DataPoint from Byte to Int due to it being signed

+++ app/src/main/java/ca/warp7/android/scouting/v5/entry/DataPoint.kt
@@ -1,5 +1,5 @@
 package ca.warp7.android.scouting.v5.entry
 
-data class DataPoint(val type: Byte, val value: Byte, val time: Byte): Iterable<Byte>{
-    override fun iterator() = byteArrayOf(type, value, time).iterator()
+data class DataPoint(val type: Int, val value: Int, val time: Int): Iterable<Byte>{
+    override fun iterator() = byteArrayOf(type.toByte(), value.toByte(), time.toByte()).iterator()
 }
